requirements.txt
aiosocks==0.2.6
amqp==5.1.1
anyio==3.5.0
asgiref==3.5.1
asttokens==2.0.5
async-timeout==4.0.2
backcall==0.2.0
billiard==3.6.4.0
black==22.3.0
celery==5.2.6
cli-exit-tools==1.2.3
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.4
dnspython==2.2.1
environs==9.5.0
executing==0.8.3
fastapi==0.77.1
h11==0.13.0
httptools==0.4.0
idna==3.3
ipython==8.3.0
jedi==0.18.1
kombu==5.2.4
lib-detect-testenv==2.0.2
marshmallow==3.15.0
matplotlib-inline==0.1.3
multiprocess==0.70.12.2
mypy-extensions==0.4.3
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.2
prompt-toolkit==3.0.29
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.9.0
Pygments==2.12.0
pyparsing==3.0.9
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
redis==4.3.1
six==1.16.0
sniffio==1.2.0
stack-data==0.2.0
starlette==0.19.1
timeout-decorator==0.5.0
tomli==2.0.1
traitlets==5.2.0
typing-extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
vine==5.0.0
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
wrapt-timeout-decorator==1.3.10
